+++ apps/lostark-hands-next/app/@component/daily-content/card.tsx
