+++ apps/lostark-hands-next/app/@component/daily-content/section.tsx
@@ -8,9 +8,10 @@ import useFilterTimerList from '@/hook/use-filter-timer-list';
 
 import { convertCalendarData } from '@/util/calendar';
 
-import DailyContentCard from '@/app/@component/daily-content-card';
 import type { ICalenderContetProps } from '@/app/@component/types';
 import { LabelLayout } from '@/client-component/label-layout';
+
+import DailyContentCard from './card';
 
 const DailyContentSection = ({ title, list }: ICalenderContetProps) => {
 	const { onOpenModal } = useModal();
