+++ apps/lostark-hands-next/app/@component/procyon-compass/card.tsx
