+++ apps/lostark-hands-next/app/page.tsx
@@ -4,18 +4,14 @@ import { getCalendarApi } from '@/service/game-contents';
 import { calendarSelector } from '@/service/game-contents/selector';
 import { getEventApi, getNoticeApi } from '@/service/news';
 
-import {
-	DailyContentSectionList,
-	DailyContentSectionListSkeleton
-} from '@/app/@component/daily-content-section-list';
 import { Event, EventSkeleton } from '@/app/@component/event';
 import { Notice, NoticeSkeleton } from '@/app/@component/notice';
 import NotificationButton from '@/app/@component/notification-button';
-import {
-	ProcyonCompassSectionList,
-	ProcyonCompassSectionListSkeleton
-} from '@/app/@component/procyon-compass-section-list';
 
+import DailyContent from './@component/daily-content';
+import DailyContentSkeleton from './@component/daily-content/skeleton';
+import ProcyonCompass from './@component/procyon-compass';
+import ProcyonCompassSkeleton from './@component/procyon-compass/skeleton';
 import ServerWrapper from './server-wrapper';
 
 /**
@@ -45,10 +41,10 @@ const Page = () => {
 	return (
 		<div className="space-y-[16px] px-[16px] pb-[16px]">
 			<ServerWrapper
-				fallback={<ProcyonCompassSectionListSkeleton />}
+				fallback={<ProcyonCompassSkeleton />}
 				apiPromise={getCalendarApi()}
 				selector={calendarSelector}
-				render={ProcyonCompassSectionList}
+				render={ProcyonCompass}
 			/>
 			<ServerWrapper
 				fallback={<EventSkeleton />}
@@ -61,10 +57,10 @@ const Page = () => {
 				render={Notice}
 			/>
 			<ServerWrapper
-				fallback={<DailyContentSectionListSkeleton />}
+				fallback={<DailyContentSkeleton />}
 				apiPromise={getCalendarApi()}
 				selector={calendarSelector}
-				render={DailyContentSectionList}
+				render={DailyContent}
 			/>
 			<NotificationButton />
 		</div>
